refactor: remove `/rmrk`
