+++ components/gallery/GalleryCard.vue
