+++ components/gallery/GalleryItem.vue
@@ -183,11 +183,10 @@ import UnlockableTag from './UnlockableTag.vue'
 import { useGalleryItem } from './useGalleryItem'
 import { GALLERY_ITEM_TABS } from '@/components/gallery/GalleryItemTabsPanel/types'
 import CollectionDetailsPopover from '@/components/collectionDetailsPopover/CollectionDetailsPopover.vue'
-import { MediaType } from '@/components/rmrk/types'
 import { usePreferencesStore } from '@/stores/preferences'
 import { exist } from '@/utils/exist'
 import { formatBalanceEmptyOnZero, formatNumber } from '@/utils/format/balance'
-import { resolveMedia } from '@/utils/gallery/media'
+import { resolveMedia, MediaType } from '@/utils/gallery/media'
 import { sanitizeIpfsUrl, toOriginalContentUrl } from '@/utils/ipfs'
 import { convertMarkdownToText } from '@/utils/markdown'
 import { generateNftImage } from '@/utils/seoImageGenerator'
