--- components/rmrk/types.ts
@@ -1,12 +1,0 @@
-// TODO: move to shared
-export enum MediaType {
-  VIDEO = 'Video',
-  MODEL = 'Model',
-  IMAGE = 'Image',
-  AUDIO = 'Audio',
-  JSON = 'Json',
-  TEXT = 'Text',
-  IFRAME = 'IFrame',
-  UNKNOWN = 'Unknown',
-  OBJECT = 'Object',
-}
