--- components/rmrk/utils.ts
@@ -1,43 +1,0 @@
-import { MediaType } from './types'
-
-// TODO: move to shared
-export const resolveMedia = (mimeType?: string): MediaType => {
-  if (!mimeType) {
-    return MediaType.UNKNOWN
-  }
-
-  if (/^application\/json/.test(mimeType)) {
-    return MediaType.MODEL
-  }
-
-  if (/^text\/html/.test(mimeType)) {
-    return MediaType.IFRAME
-  }
-
-  if (/^image\/svg\+xml/.test(mimeType)) {
-    return MediaType.IMAGE
-  }
-
-  if (/^application\/pdf/.test(mimeType)) {
-    return MediaType.OBJECT
-  }
-
-  const match = mimeType.match(/^[a-z]+/)
-
-  if (!match) {
-    return MediaType.UNKNOWN
-  }
-
-  const prefix = match[0].toUpperCase()
-
-  let result = MediaType.UNKNOWN
-
-  Object.entries(MediaType).forEach(([type, value]) => {
-    if (type === prefix) {
-      result = value
-      return
-    }
-  })
-
-  return result
-}
