+++ utils/gallery/media.ts
@@ -1,4 +1,14 @@
-import { MediaType } from '~/components/rmrk/types'
+export enum MediaType {
+  VIDEO = 'Video',
+  MODEL = 'Model',
+  IMAGE = 'Image',
+  AUDIO = 'Audio',
+  JSON = 'Json',
+  TEXT = 'Text',
+  IFRAME = 'IFrame',
+  UNKNOWN = 'Unknown',
+  OBJECT = 'Object',
+}
 
 export const mediaTypeElementSelectors: Record<
   Extract<
